Tweak build system; adds some env vars.

+++ build/build_linux_arm32.sh
@@ -5,31 +5,37 @@
 
 set -e -x
 
-mkdir  -p lib/linux_arm32/
+mkdir -p lib/linux_arm32/
 python3 src/classic_to_ctx.py
 
-rm -rf build/units_arm32
-mkdir build/units_arm32
-fpc @src/linux-arm32.cfg -B src/dss_capi.lpr
+if [[ "x${DSS_CAPI_BUILD_INC}" != "x1" ]]; then
+    rm -rf build/units_arm32 build/units_arm32_dbg
+fi
 
-rm -rf build/units_arm32
-mkdir build/units_arm32
+if [[ "x${DSS_CAPI_BUILD_DBG}" != "x1" ]]; then
+    mkdir -p build/units_arm32
+    fpc @src/linux-arm32.cfg -B src/dss_capi.lpr
+fi
+
+mkdir -p build/units_arm32_dbg
 fpc @src/linux-arm32-dbg.cfg -B src/dss_capid.lpr
 
-mkdir -p release/dss_capi/lib
-cp -R lib/linux_arm32 release/dss_capi/lib/linux_arm32
-cp -R include release/dss_capi/
-# cp -R examples release/dss_capi/
-cp LICENSE release/dss_capi/
-cp OPENDSS_LICENSE release/dss_capi/
-if [ -d "klusolvex" ]; then
-    cp klusolvex/LICENSE release/dss_capi/KLUSOLVE_LICENSE
-else  
-    cp ../klusolvex/LICENSE release/dss_capi/KLUSOLVE_LICENSE
-fi
-cd release
-tar zcf "dss_capi_${DSS_CAPI_VERSION}_linux_arm32.tar.gz" dss_capi
-cd ..
-rm -rf release/dss_capi
+if [[ "x${DSS_CAPI_BUILD_DBG}" != "x1" ]]; then
+    mkdir -p release/dss_capi/lib
+    cp -R lib/linux_arm32 release/dss_capi/lib/linux_arm32
+    cp -R include release/dss_capi/
+    # cp -R examples release/dss_capi/
+    cp LICENSE release/dss_capi/
+    cp OPENDSS_LICENSE release/dss_capi/
+    if [ -d "klusolvex" ]; then
+        cp klusolvex/LICENSE release/dss_capi/KLUSOLVE_LICENSE
+    else  
+        cp ../klusolvex/LICENSE release/dss_capi/KLUSOLVE_LICENSE
+    fi
+    cd release
+    tar zcf "dss_capi_${DSS_CAPI_VERSION}_linux_arm32.tar.gz" dss_capi
+    cd ..
+    rm -rf release/dss_capi
 
-ls release
+    ls release
+fi
